sample/app/requirements.txt
python-dotenv
flask
flask-cors
pillow
requests
termcolor
llama-index-llms-openai
